fix bugs

+++ storm/src/storm-synthesis/synthesis/Counterexample.h
@@ -12,6 +12,8 @@
 #include "storm/models/sparse/Dtmc.h"
 #include "storm/utility/Stopwatch.h"
 
+#include "storm/storage/BitVector.h"
+
 namespace storm {
     namespace synthesis {
 
@@ -51,7 +53,7 @@ namespace storm {
             void prepareDtmc(
                 storm::models::sparse::Dtmc<ValueType> const& dtmc,
                 std::vector<uint_fast64_t> const& state_map,
-                uint_fast64_t state_quant
+                size_t state_quant
                 );
             
             /*!
@@ -68,7 +70,7 @@ namespace storm {
                 ValueType formula_bound,
                 std::shared_ptr<storm::modelchecker::ExplicitQuantitativeCheckResult<ValueType> const> mdp_bounds,
                 std::vector<StateType> const& mdp_quotient_state_map,
-                uint_fast64_t state_quant
+                size_t state_quant
                 );
 
             /*!
@@ -121,7 +123,7 @@ namespace storm {
                 storm::models::sparse::StateLabeling const& labeling_subdtmc,
                 std::unordered_map<std::string,storm::models::sparse::StandardRewardModel<ValueType>> & reward_models_subdtmc,
                 std::vector<StateType> const& to_expand,
-                uint_fast64_t state_quant
+                size_t state_quant
                 );
 
             // Quotient MDP
